Changed loading logic
